+++ app/src/main/java/physics/RoamerPhysics.java
@@ -2,10 +2,10 @@ package physics;
 
 public class RoamerPhysics extends Physics {
 
-    private double baseResistance;
+    private float baseResistance;
 
-    double getHorizontalResistance(PhysicalPositional ph) {
-        double a = getBaseResistance() + ph.getSpecHorizontalResistance(this);
+    float getHorizontalResistance(PhysicalPositional ph) {
+        float a = getBaseResistance() + ph.getSpecHorizontalResistance(this);
         if (ph.horizontalCollision != 0)
             a += baseFrictionalResistance;
         if (a >= 1) return 1;
@@ -13,8 +13,8 @@ public class RoamerPhysics extends Physics {
         return a;
     }
 
-    double getVerticalResistance(PhysicalPositional ph) {
-        double a = getBaseResistance() + ph.getSpecVerticalResistance(this);
+    float getVerticalResistance(PhysicalPositional ph) {
+        float a = getBaseResistance() + ph.getSpecVerticalResistance(this);
         if (ph.verticalCollision != 0)
             a += baseFrictionalResistance;
         if (a >= 1) return 1;
@@ -22,11 +22,11 @@ public class RoamerPhysics extends Physics {
         return a;
     }
 
-    public double getBaseResistance() {
+    public float getBaseResistance() {
         return baseResistance;
     }
 
-    public void setBaseResistance(double baseResistance) {
+    public void setBaseResistance(float baseResistance) {
         if (baseResistance < 0 || baseResistance > 1)
             throw new IllegalArgumentException("baseResistance must be '0 <= x <= 1'");
         this.baseResistance = baseResistance;
